requirements.txt
Flask==3.0.3
Werkzeug==3.0.1
Flask-Mail==0.10.0
itsdangerous==2.2.0
python-dotenv==1.2.2
pymongo[srv]==4.6.3
dnspython==2.6.1
certifi==2024.2.2
requests==2.32.3
